requirements.txt
Flask
flask-cors
Pillow
requests
gunicorn
